requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==1.3.0
cycler==0.10.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dm-haiku==0.0.6
docker-pycreds==0.4.0
EasyProcess==1.1
entrypoint2==1.0
entrypoints==0.4
executing==0.8.3
flatbuffers==2.0
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gym==0.17.2
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
jax==0.3.9
jaxlib==0.3.7+cuda11.cudnn82
jedi==0.18.1
Jinja2==3.1.0
jmp==0.0.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.11.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
libclang==13.0.0
lightgbm==3.3.5
littleutils==0.2.2
llvmlite==0.33.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.2.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
notebook-shim==0.1.0
numba==0.50.0
numpy==1.22.3
oauthlib==3.2.0
obp==0.5.2
opt-einsum==3.3.0
outdated==0.2.1
packaging==21.3
pandas==1.0.4
pandas-flavor==0.2.0
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
patool==1.12
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pingouin==0.4.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pyglet==1.5.0
Pygments==2.11.2
pyieoe==0.1.1
pynvml==11.5.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.1
pytorch-lightning==1.6.3
pytrec-eval==0.5
pytz==2020.1
pyunpack==0.2.2
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.11
setproctitle==1.2.3
sharedmem==0.3.8
shortuuid==1.0.9
six==1.15.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
statsmodels==0.13.2
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
torch==1.11.0
torchmetrics==0.8.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.16
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.1.1
wrapt==1.14.0
xarray==2022.3.0
yarl==1.7.2
zipp==3.7.0
